requirements.txt
Flask
PyMuPDF
python-docx
spacy
scikit-learn
numpy
